file reorganization and additions
Reorganized file structure and worked on some encoder stuff
