+++ Software/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "signal.h": "c",
+        "pigpio.h": "c"
+    }
+}
